Refactor test to handle TLS connect in HttpTLSTest and handle proxy for HTTP/2
